review: explicit (int) array size, fold ulong-aware Math.Min into UnclesValidator
StatelessBlockTree.Prefetch: explicit `(int)length` for array creation so the
ulong→int narrowing is on display. UnclesValidator.IsKin: collapse the
header.Number ternary into Math.Min((ulong)relativeDepth, header.Number);
the previous form needed a separate (int) cast on the ulong block number.

+++ src/Nethermind/Nethermind.Consensus/Stateless/StatelessBlockTree.cs
@@ -234,7 +234,7 @@ public class StatelessBlockTree(IReadOnlyCollection<BlockHeader> headers)
     public Task<Hash256[]?> Prefetch(BlockHeader blockHeader, CancellationToken cancellationToken)
     {
         const ulong length = BlockhashCache.MaxDepth + 1;
-        Hash256[] result = new Hash256[length];
+        Hash256[] result = new Hash256[(int)length];
         result[0] = blockHeader.Hash;
         for (ulong i = 1; i < length; i++)
         {
